blog/static/tests/requirements.txt
python-dotenv
selenium
pytest-shutil